Refactored code to create MasterFilter component
