+++ frontend/src/pages/SignUp.jsx
@@ -3,6 +3,7 @@ import { useForm } from "react-hook-form";
 import { Link } from "react-router-dom";
 import axios from "axios";
 
+import MasterFilter from "@components/MasterFilter";
 import { TextField, Autocomplete } from "@mui/material";
 import { styled } from "@mui/material/styles";
 import Switch from "@mui/material/Switch";
@@ -18,7 +19,6 @@ function SignUp() {
   } = useForm();
 
   const [diplomeInput, setDiplomeInput] = useState([true, false, false]);
-  const [masterInput, setMasterInput] = useState([true, false, false]);
 
   const handleDiplomeInput = (index, type) => {
     const provDiplomeInput = [...diplomeInput];
@@ -30,23 +30,12 @@ function SignUp() {
     setDiplomeInput(provDiplomeInput);
   };
 
-  const handleMasterInput = (index, type) => {
-    const provMasterInput = [...masterInput];
-    if (index < 2 && type === "plus") {
-      provMasterInput[index + 1] = true;
-    } else if (index > 0 && type === "minus") {
-      provMasterInput[index] = false;
-    }
-    setMasterInput(provMasterInput);
-  };
-
   const onSubmit = (data) => console.warn(data);
   console.warn(errors);
 
   const [diplomeData, setDiplomeData] = useState([]);
   const [professionData, setProfessionData] = useState([]);
   const [promotionData, setPromotionData] = useState([]);
-  const [masterData, setMasterData] = useState([]);
 
   const getDiplome = () => {
     axios
@@ -85,22 +74,10 @@ function SignUp() {
       });
   };
 
-  const getMaster = () => {
-    axios
-      .get(`${import.meta.env.VITE_BACKEND_URL}/master`)
-      .then((res) => {
-        setMasterData(res.data);
-      })
-      .catch((err) => {
-        console.error(err);
-      });
-  };
-
   useEffect(() => {
     getDiplome();
     getProfession();
     getPromotion();
-    getMaster();
   }, []);
 
   const AntSwitch = styled(Switch)(({ theme }) => ({
@@ -220,7 +197,6 @@ function SignUp() {
                         <button
                           type="button"
                           disabled={index === 2}
-                          // className="flex items-center content-center"
                           className="pb-3"
                           onClick={() => handleDiplomeInput(index, "plus")}
                         >
@@ -229,7 +205,6 @@ function SignUp() {
                             color={index === 2 ? "#d3d3d3" : "black"}
                           />
                         </button>
-                        {/* )} */}
                         <button
                           type="button"
                           disabled={index === 0}
@@ -289,83 +264,7 @@ function SignUp() {
                 Champs optionnels
                 {/* BLOC MASTER ANNEE */}
               </h3>
-              {masterInput.map((masterIn, index) => {
-                if (masterIn) {
-                  return (
-                    <div className="flex flex-wrap justify-between mb-5">
-                      <Autocomplete
-                        disablePortal
-                        id="combo-box-4"
-                        options={masterData}
-                        getOptionLabel={(option) =>
-                          option.title.replace("&apos;E", "'É")
-                        }
-                        sx={{
-                          width: "55%",
-                          mb: 1.5,
-                        }}
-                        renderInput={(params) => (
-                          <TextField
-                            {...params}
-                            label={`#${index + 1} Master 2`}
-                            color="primary"
-                          />
-                        )}
-                      />
-                      <Autocomplete
-                        disablePortal
-                        id="combo-box-5"
-                        options={promotionData}
-                        getOptionLabel={(option) => option.year}
-                        sx={{
-                          width: "30%",
-                          mb: 1,
-                        }}
-                        renderInput={(params) => (
-                          <TextField
-                            {...params}
-                            label="Année"
-                            color="primary"
-                          />
-                        )}
-                      />
-                      <button
-                        type="button"
-                        disabled={index === 2}
-                        className="pb-3"
-                        onClick={() => handleMasterInput(index, "plus")}
-                      >
-                        <HiPlus
-                          size={20}
-                          color={index === 2 ? "#d3d3d3" : "black"}
-                        />
-                      </button>
-                      <button
-                        type="button"
-                        disabled={index === 0}
-                        className="pb-3"
-                        onClick={() => handleMasterInput(index, "minus")}
-                      >
-                        <HiMinus
-                          size={20}
-                          color={index === 0 ? "#d3d3d3" : "black"}
-                        />
-                      </button>
-                      {/* <div className="flex "> */}
-                      <TextField
-                        label="Université d'obtention"
-                        size="medium"
-                        sx={{
-                          width: "100%",
-                        }}
-                      />
-                      {/* </div> */}
-                    </div>
-                  );
-                }
-                return null;
-              })}
-
+              <MasterFilter />
               <TextField label="Employeur actuel" size="medium" />
               <TextField
                 label="Email professionnel"
